ref: rename module to `ifg`
