+++ .travis.yml
@@ -20,7 +20,7 @@ deploy:
   twine_version: 1.12.1
   user: __token__
   password:
-    secure: c36mLA+adviGP3FVp9vx82XRmbrQjPxtQ3KuuWnnIUKn1xDjf3YeVk6GRwYl4YP/zQNj9NLKJ4vN3LszZbssGd/BfYYlsPxf9MIM7FDgEe4UTtqHqdab2U5JG/8Ba/qa/btNHlqGKRcG4cxsYDxVxkat8H42yARyyWJlnQy0ies9GQS5qIhhReZxBUeRDlCSWcvxN4Vyvj9FSV+Yh+WujEMBR6pCyUyLN/O98S36O91r/VKDIi2tfSYgj7NfzL5aB4ve9o4Telurboxxtj1/OUdaJ9HNNduuNXDo4Ki4Lgf7258XuZB+zcYgIv7TBAbN/ef7r0bHpXSWDex52ZJyCAMqWhT/Kwgl8V+BnAMHLypVkGG9BJarjyCYisjYYn+5ISCestaqOacTt2WBbOT2rui0IIXTVrftNxKMNLDqV6ja9VIY31IXNg3fQBsCm/hkU0mU2/9D7+nSb1pthm1aR+muDmBGqLDpKTtL30FLJOj5y9BYZIPU9A4zKjn89Cjg8crALXrR2FCUujcEf82HocjOl2Rk64rGqs6MCnSO2xz0ltgO7YRMWtfEuHk02iZqMo9m3xZhpLeDjarEl2d0KykPnkMUwaqHf8xw+PBhhb4tdQoNkCeOUNRIL0Rhul1WT8e6ci1WFi7dzi2fxw65olGFhgVF5AGtRvVRJvDbsPs=
+    secure: DM1F0tApae9m7G+8xFQf9szjwalhLY5fGcyPFpNMR1nQ8T2JCQRi549lUUTL2/kgWrsfHyfWklmc9gxc7bk1XkeHsHX3vhKRPAXBhG4bZAWyPJBbx81qCpZ1NRqFiwISjFkE0noceBYNd0CuJTuPvoRcOJ9fIeVtKk7SHhT/8okgJYZX8BeE+RUUy3HPRSNqQL7xNQfJIOWZ4mlhvJwOgUB66IjUKTtxdEZL4WQ47+G5YhxpcUP6QLlbbhi8J4KJsuw7XftLKhxSzw8Mh53Xq8M6b+av3S/ysNzKumrogGlPkz58pun/QHuAA8KRf7uoVarjY99yiN7wFzSmrGwtJB5kkLl9h1TkhqcSm87KllGXa0GsdUwMyEdBCM+ByA3CmgEN6tjj3YdDUm0AY66LPHMSG7OM0+T3Fuv4kIIMr33LCIVzxLJEMGXjWZu1M4Vxzw8aVwH97qnI+x+zwUzA0K/ha59v/GYMt7O2oGU+KQy0MwXn+1b17MY0756uAOLSJnblT0CcUUVaOtgENY1cPvFFIcZ4mgiG/xw7NFQNEjH2Q6Hq6xaYFMkO6Xny2le2P/T/x7qlBh9pS0ftvDGxB8mG8hiTqt/l7t4iDMU9YaztJBKnsgu90tYFh52GvM9kXzxTucOEWbPz0iliEGBPX991EG5p0Ib8Qof8Kgy0dfg=
   on:
     branch: master
   skip_existing: true
