+++ docs/source/conf.py
@@ -13,7 +13,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../../'))
-from ifg_py.metadata import __version__     # noqa: E402
+from ifg.metadata import __version__     # noqa: E402
 
 
 # -- Project information -----------------------------------------------------
