+++ ifg/__init__.py
@@ -1,0 +1,9 @@
+from ifg.calculator import IfgCalculator
+from ifg.units_converter import SiAtomicConverter, get_metal_specific_volume
+
+from ifg.metadata import __version__
+
+__all__ = [
+    'IfgCalculator', 'SiAtomicConverter',
+    'get_metal_specific_volume', '__version__'
+]
