+++ ifg/metadata.py
