--- ifg_py/__init__.py
@@ -1,9 +1,0 @@
-from ifg_py.calculator import IfgCalculator
-from ifg_py.units_converter import SiAtomicConverter, get_metal_specific_volume
-
-from ifg_py.metadata import __version__
-
-__all__ = [
-    'IfgCalculator', 'SiAtomicConverter',
-    'get_metal_specific_volume', '__version__'
-]
