+++ tox.ini
@@ -10,7 +10,7 @@ envlist = python2.7, python3.5, python3.6, python3.7, docs, flake8
 deps =
     -rrequirements.txt
 commands =
-    python -m ifg_py.examples
+    python -m ifg.examples
     pytest tests/unit
 
 # Documentation
@@ -28,4 +28,4 @@ deps =
     -rrequirements.txt
     flake8
 commands =
-    flake8 ifg_py
+    flake8 ifg
